살려줘

+++ backend/matchpet/src/main/java/com/matchpet/domain/manager/entity/ManagerProfile.java
@@ -1,5 +1,7 @@
 package com.matchpet.domain.manager.entity;
 
+import java.math.BigDecimal;
+
 import com.matchpet.domain.match.Enums.Level3;
 import jakarta.persistence.*;
 import lombok.Getter; import lombok.Setter;
@@ -20,7 +22,7 @@ public class ManagerProfile {
   private Level3 elderlyExpLevel;   // LOW/MID/HIGH
 
   @Column(precision = 3, scale = 2)
-  private Double reliabilityScore;  // 0.00 ~ 1.00
+  private BigDecimal reliabilityScore;  // 0.00 ~ 1.00
 
   @Column(columnDefinition = "json")
   private String animalSkillTags;   // ["high_energy_handling","device_friendly"]
